// backends/api-beans/src/.generated/graphql.ts
export type Maybe<T> = T extends PromiseLike<infer U> ? Promise<U | null> : T | null;
export type InputMaybe<T> = T extends PromiseLike<infer U> ? Promise<U | null> : T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar.This scalar is serialized to a string in ISO 8601 format and parsed from a string in ISO 8601 format. */
  DateTimeISO: { input: Date; output: Date; }
  /** A field whose value is a valid decimal degrees latitude number (53.471): https://en.wikipedia.org/wiki/Latitude */
  Latitude: { input: number; output: number; }
  /** A field whose value is a valid decimal degrees longitude number (53.471): https://en.wikipedia.org/wiki/Longitude */
  Longitude: { input: number; output: number; }
  /** A field whose value conforms to the standard URL format as specified in RFC3986: https://www.ietf.org/rfc/rfc3986.txt. */
  URL: { input: string; output: string; }
  /** A field whose value is a generic Universally Unique Identifier: https://en.wikipedia.org/wiki/Universally_unique_identifier. */
  UUID: { input: string; output: string; }
};

export type IBaseBean = {
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  originId?: Maybe<Scalars['UUID']['output']>;
  processingMethodId?: Maybe<Scalars['UUID']['output']>;
  roastLevelId?: Maybe<Scalars['UUID']['output']>;
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IBaseBeanProcessingMethod = {
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IBaseBeanRoastLevel = {
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IBaseFlavorProfile = {
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  label: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
  value: Scalars['String']['output'];
};

export type IBaseOrigin = {
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  latitude?: Maybe<Scalars['Latitude']['output']>;
  longitude?: Maybe<Scalars['Longitude']['output']>;
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IBasePack = {
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};
// ...
export type IBaseTasting = {
  acidity: Scalars['Int']['output'];
  conductedAt: Scalars['DateTimeISO']['output'];
  createdAt: Scalars['DateTimeISO']['output'];
  description: Scalars['String']['output'];
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  richness: Scalars['Int']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IBean = IBaseBean & {
  __typename?: 'Bean';
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  flavorProfiles?: Maybe<Array<Maybe<IFlavorProfile>>>;
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  origin?: Maybe<IOrigin>;
  originId?: Maybe<Scalars['UUID']['output']>;
  packs?: Maybe<Array<Maybe<IPack>>>;
  processingMethod?: Maybe<IProcessingMethod>;
  processingMethodId?: Maybe<Scalars['UUID']['output']>;
  roastLevel?: Maybe<IRoastLevel>;
  roastLevelId?: Maybe<Scalars['UUID']['output']>;
  tastings?: Maybe<Array<Maybe<ITasting>>>;
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IBeanResults = IBean | INotFoundError;

export type IConflictError = IGraphQlError & {
  __typename?: 'ConflictError';
  code: Scalars['String']['output'];
  message: Scalars['String']['output'];
};

export type ICreateBeanInput = {
  description?: InputMaybe<Scalars['String']['input']>;
  name: Scalars['String']['input'];
  origin?: InputMaybe<ICreateOriginInput>;
};

export type ICreateBeanResults = IBean | IConflictError;

export type ICreateOriginInput = {
  description?: InputMaybe<Scalars['String']['input']>;
  id?: InputMaybe<Scalars['UUID']['input']>;
  latitude?: InputMaybe<Scalars['Latitude']['input']>;
  longitude?: InputMaybe<Scalars['Longitude']['input']>;
  name: Scalars['String']['input'];
};

export type ICreateOriginResults = IConflictError | IOrigin;

export type IFlavorProfile = IBaseFlavorProfile & {
  __typename?: 'FlavorProfile';
  beans?: Maybe<Array<IBean>>;
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  label: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
  value: Scalars['String']['output'];
};
// ...
export type IOrigin = IBaseOrigin & {
  __typename?: 'Origin';
  beans?: Maybe<Array<IBean>>;
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  latitude?: Maybe<Scalars['Latitude']['output']>;
  longitude?: Maybe<Scalars['Longitude']['output']>;
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IOriginResults = INotFoundError | IOrigin;

export type IOriginWhereInput = {
  name?: InputMaybe<Scalars['String']['input']>;
  q?: InputMaybe<Scalars['String']['input']>;
};

export type IPack = IBasePack & {
  __typename?: 'Pack';
  beans?: Maybe<Array<IBean>>;
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  purchases?: Maybe<Array<IPurchase>>;
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IProcessingMethod = IBaseBeanProcessingMethod & {
  __typename?: 'ProcessingMethod';
  beans?: Maybe<Array<IBean>>;
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IPurchase = IBasePurchase & {
  __typename?: 'Purchase';
  capacity: Scalars['Int']['output'];
  createdAt: Scalars['DateTimeISO']['output'];
  id: Scalars['UUID']['output'];
  pack?: Maybe<IPack>;
  packId: Scalars['UUID']['output'];
  price: Scalars['Int']['output'];
  purchasedAt: Scalars['DateTimeISO']['output'];
  shop?: Maybe<IShop>;
  shopId: Scalars['UUID']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};
// ...
export type IRoastLevel = IBaseBeanRoastLevel & {
  __typename?: 'RoastLevel';
  beans?: Maybe<Array<IBean>>;
  createdAt: Scalars['DateTimeISO']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type IShop = IBaseShop & {
  __typename?: 'Shop';
  createdAt: Scalars['DateTimeISO']['output'];
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};

export type ITasting = IBaseTasting & {
  __typename?: 'Tasting';
  acidity: Scalars['Int']['output'];
  beans?: Maybe<Array<IBean>>;
  conductedAt: Scalars['DateTimeISO']['output'];
  createdAt: Scalars['DateTimeISO']['output'];
  description: Scalars['String']['output'];
  id: Scalars['UUID']['output'];
  name: Scalars['String']['output'];
  richness: Scalars['Int']['output'];
  updatedAt: Scalars['DateTimeISO']['output'];
};